tests/demultiplex/test_parse_excel.py
from pathlib import Path
import yaml

from delt_core.demultiplex import experiment_from_excel, whitelists_from_excel, catalog_from_excel, config_from_excel, structure_from_excel

path = Path('templates/library.xlsx')
save_config = config_path = Path('templates/config.yaml')

experiment_from_excel(path)
catalog_from_excel(path)
whitelists_from_excel(path)
structure_from_excel(path)
config = config_from_excel(path)

with open(save_config, 'w') as f:
    yaml.dump(config, f, sort_keys=False, indent=2)